% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticSynthFunctions.R
\name{ElasticSynth}
\alias{ElasticSynth}
\title{ElasticSynth}
\usage{
ElasticSynth(PredictorMatrix, OutcomeMatrix, treated, pre, post,
  lambda_grid = c(seq(from = 0.001, to = 0.1, by = 0.01), seq(from = 0.2, to =
  100, by = 0.1), seq(from = 200, to = 50000, by = 100)), a_grid = seq(from =
  0.1, to = 0.9, by = 0.01), start_month, end_month, time_unit,
  penaltyMat = matrix(rep(1, ncol(OutcomeMatrix) * (ncol(OutcomeMatrix) - 1)),
  ncol = ncol(OutcomeMatrix), nrow = ncol(OutcomeMatrix)),
  max_number_units = ncol(OutcomeMatrix), lower_limit_weights = 0)
}
\arguments{
\item{PredictorMatrix}{a p x n matrix where each row is a seperate control variable to be included in the analysis for each unit}

\item{OutcomeMatrix}{a t x n matrix where each row is the outcome variable of interest at time t for each unit}

\item{treated}{the column index for the treated unit in both PredictorMatrix and OutcomeMatrix. For multiple treated units, see 'ElasticSynthRun'}

\item{pre}{a vector of row indices indicating the pre period (Ex. 1:40). This is the period over which the algorithm will optimize weights.}

\item{post}{a vector of row indices indicating the post period (Ex. 41:50)}

\item{lambda_grid}{glmnet param - a vector of values for lambda to fitted at each alpha value during cross validation. For more information, see glmnet}

\item{start_month}{the start month of the case study (first pre-period). This is used for plotting}

\item{end_month}{the end month of the case study (last post-period). This is used for plotting}

\item{time_unit}{The gap between observations in the case study. No default. A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by an integer and a space, or followed by "s". Ex '3 months' or 'quarter'}

\item{penaltyMat}{glmnet param - (optional) a n-1 x n matrix of additional penalties placed upon glmnet for selecting certain units. Useful in the case of multiple treated units. Each column of the matrix represents the penalties to be placed on the control units. The treated unit is excluded from the column. See glmnet param penalty.factor for more details. Default is no additional penalties on any unit.}

\item{max_number_units}{glmnet param - The maximum number of units glmnet will select. Useful for forcing parsimonious models. Default allows all units to receive weight.}

\item{lower_limit_weights}{The lower limit value of weight that can be placed on any unit. Default is zero (non-negative constraint). Change if desired, but be wary of overfitting.}
}
\value{
list containing output weights for treated unit (w_final), the actual outcome values (Y_true), the fitted outcome values (Y_elast), optimal value of lambda (lambda_opt), optimal value of alpha, (alpha_opt), a dataframe of the results of placebo test (placebo_frame), a plot of the path of the treated vs. actual unit (path.plot), and a plot of the placebo test results for the treated unit.
}
\description{
A synthetic controls method for comparitive case studies
that utilizes elastic net regression -- glmnet -- to assign unit weights. 
Utilizes a cross-validation mechanism for selecting optimal L1 and L2 regularization parameters, as well as lambda.
}
