% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticSynthFunctions.R
\name{ElasticSynthCovariates}
\alias{ElasticSynthCovariates}
\title{ElasticSynthCovariates}
\usage{
ElasticSynthCovariates(treatedUnits, weights, covariate, pre, post, start_month,
  end_month, time_unit, storage_folder)
}
\arguments{
\item{treatedUnits}{a list of column indices of treated units in the covariate matrix}

\item{weights}{either a list or a list of lists of the weights generated from ElasticSynth or ElasticSynthRun}

\item{pre}{a vector of pre periods. Since no optimization is happening in this step, enter the minimum pre period and the maximum pre period}

\item{post}{a vector of post periods. Since no optimization is happening in this step, enter the minimum post period and the maximum post period}

\item{start_month}{the start month of the case study (first pre-period). This is used for plotting}

\item{end_month}{the end month of the case study (last post-period). This is used for plotting}

\item{time_unit}{The gap between observations in the case study. No default. A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by an integer and a space, or followed by "s". Ex '3 months' or 'quarter'}

\item{storage_folder}{The directory where graph outputs should be stored}
}
\value{
Doesn't return an object, but will output path plot and gap plot to a directory
}
\description{
A function that uses the weights generated from ElasticSynth or ElasticSynthRun to explore covariates. In other words, this function
asks "in a synthetic unit similar to unit x in terms of outcome y, how does the synthetic unit's variable z compare to the actual unit?"
By finding large differences between the synthetic and actual unit, you may find what is driving the change in the intervention.
}
