% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_public_crime_data.R
\name{generate_public_crime_data}
\alias{generate_public_crime_data}
\title{Generate public crime data for use with ElasticSynth}
\usage{
generate_public_crime_data()
}
\description{
Generate public crime data for use with ElasticSynth
}
