% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElasticSynth.R
\name{ElasticSynth}
\alias{ElasticSynth}
\title{ElasticSynth - a class for running synthetic controls with elastic net regularization}
\description{
Model is initiated with a list of parameters most which include treated_units and donor_units, two dataframes that 
contain time periods, units, and counts of outcomes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ElasticSynth$new()}}
\item \href{#method-cv_treated}{\code{ElasticSynth$cv_treated()}}
\item \href{#method-cv_untreated}{\code{ElasticSynth$cv_untreated()}}
\item \href{#method-cv}{\code{ElasticSynth$cv()}}
\item \href{#method-generate_weights}{\code{ElasticSynth$generate_weights()}}
\item \href{#method-fit}{\code{ElasticSynth$fit()}}
\item \href{#method-plot}{\code{ElasticSynth$plot()}}
\item \href{#method-generate_hyndman_folds}{\code{ElasticSynth$generate_hyndman_folds()}}
\item \href{#method-long_to_wide}{\code{ElasticSynth$long_to_wide()}}
\item \href{#method-long_to_wide_weights}{\code{ElasticSynth$long_to_wide_weights()}}
\item \href{#method-clone}{\code{ElasticSynth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$new(
  treated_units,
  donor_units,
  measure_vars,
  measure_col,
  unit_col,
  time_col,
  value_col,
  pre_list,
  post_list,
  start_cv_list,
  end_cv_list,
  cv_step_list,
  start_month_list,
  end_month_list,
  time_unit_list,
  alphas = NULL,
  lambdas = NULL,
  lower_limit_weights,
  upper_limit_weights,
  placebo,
  verbose,
  max_pre_month,
  end_month,
  start_month
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treated_units}}{Treated units data.frame}

\item{\code{donor_units}}{Donor units data.frame}

\item{\code{measure_vars}}{Outcome measures upon which to jointly fit a synthetic control}

\item{\code{measure_col}}{Column name for measure_vars}

\item{\code{unit_col}}{Column name for unit information}

\item{\code{time_col}}{Column name for time information (in form 1:max(time))}

\item{\code{pre_list}}{A list of lists or vectors representing pre periods across all outcomes}

\item{\code{post_list}}{A list of lists or vectors representing post periods across all outcomes}

\item{\code{start_cv_list}}{A list indicating where to start CV procedure in time}

\item{\code{end_cv_list}}{A list indicating where to end CV procedure in time}

\item{\code{time_unit_list}}{A list indicating the time between time periods}

\item{\code{alphas}}{vector of alphas to use for cross validation - for more information see glment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv_treated"></a>}}
\if{latex}{\out{\hypertarget{method-cv_treated}{}}}
\subsection{Method \code{cv_treated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$cv_treated()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv_untreated"></a>}}
\if{latex}{\out{\hypertarget{method-cv_untreated}{}}}
\subsection{Method \code{cv_untreated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$cv_untreated()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv"></a>}}
\if{latex}{\out{\hypertarget{method-cv}{}}}
\subsection{Method \code{cv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$cv(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_weights"></a>}}
\if{latex}{\out{\hypertarget{method-generate_weights}{}}}
\subsection{Method \code{generate_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$generate_weights(cv_results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$fit(weights, cv_results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$plot(series_frame)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_hyndman_folds"></a>}}
\if{latex}{\out{\hypertarget{method-generate_hyndman_folds}{}}}
\subsection{Method \code{generate_hyndman_folds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$generate_hyndman_folds(start_cv, end_cv, cv_step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_to_wide"></a>}}
\if{latex}{\out{\hypertarget{method-long_to_wide}{}}}
\subsection{Method \code{long_to_wide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$long_to_wide(dt, time_col, unit_col, value_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_col}}{Column name for time information (in form 1:max(time))}

\item{\code{unit_col}}{Column name for unit information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_to_wide_weights"></a>}}
\if{latex}{\out{\hypertarget{method-long_to_wide_weights}{}}}
\subsection{Method \code{long_to_wide_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$long_to_wide_weights(
  dt,
  pre_list,
  time_col,
  unit_col,
  measure_col,
  value_col,
  fitted_vars
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pre_list}}{A list of lists or vectors representing pre periods across all outcomes}

\item{\code{time_col}}{Column name for time information (in form 1:max(time))}

\item{\code{unit_col}}{Column name for unit information}

\item{\code{measure_col}}{Column name for measure_vars}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElasticSynth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
